feat: add UserNotFoundError

+++ src/application/delete_user/delete_user_command_handler.py
@@ -1,4 +1,7 @@
+from uuid import UUID
+
 from src.application.delete_user.delete_user_command import DeleteUserCommand
+from src.domain.user_not_found_error import UserNotFoundError
 from src.domain.user_repository import UserRepository
 
 
@@ -7,5 +10,7 @@ class DeleteUserCommandHandler:
         self._repository = repository
 
     def execute(self, command: DeleteUserCommand) -> None:
-        user = self._repository.find(user_id=command.id)
+        user = self._repository.find(user_id=UUID(command.id))
+        if user is None:
+            raise UserNotFoundError(command.id)
         self._repository.delete(user.id)
